Update search.py
This PR should allow for easier retrieval of historical posts by toggling the ability to search existing threads from deleted users to ON, and if I'm right (I'm probably not) then it should also replace all spaces in search queries with wildcards. Currently users have to insert their own wildcard (%) to not only search for the exact string they typed. But this isn't explained, and is thus functionally useless to the overwhelming majority of users.

Please approve asap and give badge thanks ily